feat: notes getting added from sidebar
